graphKpop/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
neo4j==5.8.1
psutil==5.9.5
praw==7.7.0

# linters
mypy==1.3.0
black==23.3.0
pylint==2.17.4
flake8==6.0.0
types-requests==2.31.0.0
pandas-stubs==2.0.1.230501
types-pytz==2023.3.0.0
types-psutil==5.9.5.12
types-openpyxl==3.1.0.8